refactor: Modal 컴파운드 패턴 적용 #45
