--- src/components/ModalActions/index.tsx
@@ -1,16 +1,0 @@
-import { StrictPropsWithChildren } from "@/types";
-
-import { Container } from "./style";
-
-type Direction = "row" | "col";
-
-export interface ModalActionsProps {
-  direction?: Direction; // 정렬 방향
-}
-
-export default function ModalActions({
-  direction = "row",
-  children,
-}: StrictPropsWithChildren<ModalActionsProps>) {
-  return <Container direction={direction}>{children}</Container>;
-}
