--- src/components/ModalContent/index.tsx
@@ -1,7 +1,0 @@
-import { StrictPropsWithChildren } from "@/types";
-
-import { Container } from "./style";
-
-export default function ModalContent({ children }: StrictPropsWithChildren) {
-  return <Container>{children}</Container>;
-}
