--- src/components/ModalContent/style.ts
@@ -1,5 +1,0 @@
-import styled from "@emotion/styled";
-
-export const Container = styled.div`
-  padding: 16px 24px 12px 24px;
-`;
